requirements.txt
fastapi
uvicorn
notion-client
langchain
openai
python-dotenv
langchain-openai